requirements.txt
django == 5.1.1
pip == 24.1.2
beautifulsoup4 == 4.12.3
requests == 2.32.3
djangorestframework == 3.15.2